requirements.txt
openai==1.34.0
numpy==1.26.4
python-dotenv==1.0.1
tiktoken==0.7.0
opencv-python==4.10.0.82
selenium==4.21.0
faiss-cpu==1.8.0